travel_planner/requirements.txt
Flask
requests
beautifulsoup4
groq
httpx
python-dotenv